requirements.txt
amqp==5.1.1
asgiref==3.5.1
async-timeout==4.0.2
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==3.2.13
django-celery-beat==2.2.1
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.3.0
django-environ==0.8.1
django-redis==5.2.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.1.0
pluggy==1.0.0
prompt-toolkit==3.0.29
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.8
pytest==7.1.2
pytest-django==4.5.2
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.2.2
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
